+++ ebcdic/src/commonMain/kotlin/com/lppedd/kotlinx/charset/EbcdicDbcsDecoder.kt
@@ -30,20 +30,18 @@ internal class EbcdicDbcsDecoder(
       val b1 = bytes[i++].toInt() and 0xFF /* to unsigned */
 
       when (b1) {
-        Ebcdic.SO -> if (mode != Ebcdic.SBCS) {
+        Ebcdic.SO.toInt() -> if (mode != Ebcdic.SBCS) {
           val c = replOrThrow("Malformed input")
           sb.append(c)
         } else {
           mode = Ebcdic.DBCS
         }
-
-        Ebcdic.SI -> if (mode != Ebcdic.DBCS) {
+        Ebcdic.SI.toInt() -> if (mode != Ebcdic.DBCS) {
           val c = replOrThrow("Malformed input")
           sb.append(c)
         } else {
           mode = Ebcdic.SBCS
         }
-
         else -> {
           var c: Char
